requirements.txt
Flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
supabase>=1.0.0
pytz>=2023.3
Pillow>=10.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
urllib3>=1.26.0
gunicorn>=21.2.0